--- a/examples/local_module.py
+++ b/examples/local_module.py
@@ -1,4 +1,7 @@
 """
+Local module
+============
+
 A trivial local module to provide a value for plot_exp.py.
 """