--- a/setup.cfg
+++ b/setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-addopts = --cov-report= --cov=sphinx_gallery --durations=5 -ra
+addopts = --cov-report= --cov=sphinx_gallery --durations=5 -ra --showlocals
 python_files = tests/*.py
 norecursedirs = build _build auto_examples gen_modules
                 sphinx_gallery/tests/tinybuild