… doc/configuration.rst