This writes a few files with a `.new` extension and replaces the old one only if necessary (e.g., `.py`, `.rst` files) to avoid Sphinx rebuilding them. Most of the added lines are new tests adapted from #466.

@NicolasHug @jklymak can you see if it works for you?

Closes #449
Closes #446
Closes #400
Closes #395
Closes #394