@jklymak 2019-02-25
On sphinx-gallery: 
  - first call: 11.44 s
  - second call: 4.4s
  - change one tutorial: 4.8 s

Checking matplotlib; update in about 30 minutes :wink: