@jklymak 2019-02-25
In matplotlib, after running `git clean -xdf` in the `doc` directory, I get an exception when I run 
`time make SPHINXOPTS= html`:

```
Exception occurred:
  File "/Users/jklymak/sphinx-gallery/sphinx_gallery/utils.py", line 98, in get_md5sum
    with open(src_file, 'rb') as src_data:
FileNotFoundError: [Errno 2] No such file or directory: '/Users/jklymak/matplotlib/doc/api/_as_gen/matplotlib.pyplot.Subplot.examples.new'
```

Note that there *are* lots of *.new files in that directory, so not entirely clear what didn't work...