@NicolasHug 2019-02-25
On sklearn sphinx tells me that it's seeing about 60 changed files (instead of hundreds) and the second build runs in a few seconds.
In other words: looks like it  works \o/