@jklymak 2019-02-25
I wonder if the above issue is something funky about the cases:

```
ls api/_as_gen/matplotlib.pyplot.Subplot*
matplotlib.pyplot.Subplot.examples              matplotlib.pyplot.subplot_tool.rst
matplotlib.pyplot.subplot.rst                   matplotlib.pyplot.subplots.examples
matplotlib.pyplot.subplot2grid.examples         matplotlib.pyplot.subplots.rst
matplotlib.pyplot.subplot2grid.rst              matplotlib.pyplot.subplots_adjust.examples.new
matplotlib.pyplot.subplot_tool.examples         matplotlib.pyplot.subplots_adjust.rst
```

Note that `Subplot.examples` corresponds with `subplots.rst`.