@larsoner 2019-02-25
@jklymak I can't replicate that error in building the `matplotlib` doc. First build I get `real	5m7.390s` and second `real	1m1.422s`, no errors. And this is what I see on my file system:
```
$ ls api/_as_gen/matplotlib.pyplot.Subplot*
api/_as_gen/matplotlib.pyplot.Subplot.examples
```
I don't know why you see any other files since the case does not match. Are you on Windows? Maybe there is some file-case-insensitivity problem...?

Is it possible you are on some broken `matplotlib` branch and not clean latest `master`?

Looking at the code, I'm not sure how it's possible you see this error. Can you paste the full traceback? Is it in `finalize_backreferences`?

> Note that Subplot.examples corresponds with subplots.rst.

Regardless of where `Subplot.examples` is included in the RST files, the `Subplot.examples` filee should be first written as an empty file, then with meaningful content as `Subplot.examples.new`, then renamed (if md5sum differs).

Nonetheless can you try the latest commit?