@larsoner 2019-02-25
Ready for review/merge from my end. @choldgraf feel free to have a look if you have time.