@jklymak 2019-02-25
I'm on a mac, so there are definite case issues (`subplot` = `Subplot`)....