@jklymak 2019-02-25
Works now, so long as I don't fail on warning for matplotlib.  OTOH:

### This PR:

 - 563s for first build,
 - 85s for second build no changes
 - 250s  changed one file in `tutorials`; Most of time is spent in `reading sources... api/_as_gen/*`

### Master:

 - 122s second build no changes; lots of time spent in `reading sources... `
 - 258s for one file changed in `tutorials`; Most of time is spent in `reading sources... api/_as_gen/*`

So, this doesn't particularly speed up incremental building (8s has to be in the noise compared to other stuff I'm doing on my machine)...