@larsoner 2019-02-25
Maybe your `sphinx` is old / bad at detecting changes? For me I get:

* 311s for first build
* 61s for second build no changes (almost all time in writing output)
* 62s for third build after adding 1 char to the intro paragraph of `tutorials/introductory/images.py` (which SG reports as being rebuilt, and I can verify to be the case viewing the output)

I can confirm this behavior makes sense by:
```
$ ls -alt api/_as_gen/ | more
total 9356
-rw-r--r-- 1 larsoner larsoner  69798 Feb 25 16:43 matplotlib.pyplot.figure.examples
-rw-r--r-- 1 larsoner larsoner   8924 Feb 25 16:43 matplotlib.pyplot.colorbar.examples
-rw-r--r-- 1 larsoner larsoner   9462 Feb 25 16:43 matplotlib.pyplot.imshow.examples
-rw-r--r-- 1 larsoner larsoner    890 Feb 25 16:43 matplotlib.image.imread.examples
-rw-r--r-- 1 larsoner larsoner   2076 Feb 25 16:43 matplotlib.pyplot.hist.examples
-rw-r--r-- 1 larsoner larsoner      0 Feb 25 16:32 mpl_toolkits.mplot3d.Axes3D.text.examples
-rw-r--r-- 1 larsoner larsoner      0 Feb 25 16:32 mpl_toolkits.mplot3d.Axes3D.quiver.examples
-rw-r--r-- 1 larsoner larsoner      0 Feb 25 16:32 mpl_toolkits.mplot3d.Axes3D.bar.examples
...
```
You can clearly see only five `.examples` were changed (the ones used by that example) when I ran it the for that third build.

Moreover, if instead of changing the *introductory paragraph* or title (which SG turns into a tooltip in the `.examples`, forcing it to change on the run) I change something in the *main body text* of the script (e.g., `specific argument` to `specificc argument`) there are no `.examples` files changed at all, so it has even less of an effect. But even changing a few of these `.examples` files should not be catastrophic.

Did you change some other example? If so, maybe it has worse consequences and we can track down why. Can you see if you can replicate what I see by just changing `tutorials/introductory/images.py`? If it builds fast for you in that case, I'm curious which example you changed that made it take so much time in the reading sources stage. If it does not build quickly for you, perhaps try updating Sphinx to see if it helps.