@jklymak 2019-02-25
I just updated to sphinx 1.8.4.  This is on a Mac.  

### This PR
- 67s second run no changes...  reading sources very fast!
- 202s: one source changed (added a character to `tutorials/intermediate/artists.py`):  [reading sources...] goes very slowly. See the output below...

```
computation time summary:
building [mo]: targets for 0 po files that are out of date
building [html]: targets for 2264 source files that are out of date
updating environment: [config changed] 2264 added, 4 changed, 0 removed
reading sources... [ 52%] api/_as_gen/mpl_toolkits.axes_grid1.axes_divider.make_axes_area_auto_adjustablreading sources... [ 57%] api/_as_gen/mpl_toolkits.axisartist.grid_helper_curvelinear.FixedAxisArtistHelreading sources... [ 57%] api/_as_gen/mpl_toolkits.axisartist.grid_helper_curvelinear.FloatingAxisArtistreading sources... [ 57%] api/_as_gen/mpl_toolkits.axisartist.grid_helper_curvelinear.GridHelperCurveLinreading sources... [100%] users/whats_new
```

- third run no changes; note the `0 added` (and reading sources goes extremely quickly)

```
computation time summary:
building [mo]: targets for 0 po files that are out of date
building [html]: targets for 2264 source files that are out of date
updating environment: [] 0 added, 35 changed, 0 removed
reading sources... [100%] tutorials/toolkits/sg_execution_times
```